# Conflicts:
#	FrontEnd/app/src/main/res/values/strings.xml

+++ FrontEnd/app/src/main/java/com/ssafy/closetory/dto/ClothItemDto.kt
@@ -1,3 +1,3 @@
 package com.ssafy.closetory.dto
 
-data class ClothItemDto(val clothesId: Long, val photoUrl: String)
+data class ClothItemDto(val clothesId: Int, val photoUrl: String)
